Renamed video.h/.cpp to VideoSystem.h/.cpp
